requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
django-cors-headers
pillow
psycopg2
djangorestframework-simplejwt